Cleanup. The path does no longer contain the starting position. Scripts for testing.

+++ test/ScenarioLoader.h
@@ -23,16 +23,16 @@ class ScenarioLoader;
 
 class Experiment {
 public:
-	Experiment(int sx,int sy,int gx,int gy,int b, float d, string m)
+	Experiment(int sx,int sy,int gx,int gy,int b, double d, string m)
     :startx(sx),starty(sy),goalx(gx),goaly(gy),scaleX(kNoScaling),scaleY(kNoScaling),bucket(b),distance(d),map(m){}
-	Experiment(int sx,int sy,int gx,int gy,int sizeX, int sizeY,int b, float d, string m)
+	Experiment(int sx,int sy,int gx,int gy,int sizeX, int sizeY,int b, double d, string m)
     :startx(sx),starty(sy),goalx(gx),goaly(gy),scaleX(sizeX),scaleY(sizeY),bucket(b),distance(d),map(m){}
 	int GetStartX() const {return startx;}
 	int GetStartY() const {return starty;}
 	int GetGoalX() const {return goalx;}
 	int GetGoalY() const {return goaly;}
 	int GetBucket() const {return bucket;}
-	float GetDistance() const {return distance;}
+	double GetDistance() const {return distance;}
 	//void GetMapName(char* mymap) const {strcpy(mymap,map.c_str());}
 	const char *GetMapName() const { return map.c_str(); }
 	int GetXScale() const {return scaleX;}
@@ -44,7 +44,7 @@ private:
 	int scaleX;
 	int scaleY;
 	int bucket;
-	float distance;
+	double distance;
 	string map;
 };
